+++ src/main/java/carracing/domain/NumberOfAttempts.java
@@ -1,4 +1,4 @@
-package carracing.model;
+package carracing.domain;
 
 import static carracing.RacingGameSetupValidator.isPositiveInteger;
