+++ src/main/java/carracing/domain/NumberOfCars.java
@@ -1,4 +1,4 @@
-package carracing.model;
+package carracing.domain;
 
 import static carracing.RacingGameSetupValidator.isPositiveInteger;
