+++ src/main/java/carracing/domain/car/RaceRecordOfCar.java
@@ -1,4 +1,4 @@
-package carracing.model.car;
+package carracing.domain.car;
 
 import java.util.List;
