+++ src/main/java/carracing/domain/car/strategy/MovingStrategy.java
@@ -1,4 +1,4 @@
-package carracing.model.car.strategy;
+package carracing.domain.car.strategy;
 
 @FunctionalInterface
 public interface MovingStrategy {
