+++ src/test/java/carracing/domain/NumberOfCarsTest.java
@@ -1,4 +1,4 @@
-package carracing.model;
+package carracing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
