+++ src/test/java/carracing/domain/PositionOfCarTest.java
@@ -1,5 +1,6 @@
-package carracing.model;
+package carracing.domain;
 
+import carracing.domain.car.PositionOfCar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
