+++ src/test/java/carracing/domain/car/RaceRecordOfCarTest.java
@@ -1,4 +1,4 @@
-package carracing.model.car;
+package carracing.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
